requirements.txt
Flask==0.10.1
Jinja2==2.7.3
MarkupSafe==0.23
Pillow==2.6.1
RPi.GPIO==0.6.3
RTIMULib==7.2.1
Werkzeug==0.9.6
argparse==1.2.1
blinker==1.3
chardet==2.3.0
colorama==0.3.2
google-api-python-client==1.6.2
gpiozero==1.3.2
html5lib==0.999
httplib2==0.10.3
itsdangerous==0.24
lxkeymap==0.1
mcpi==0.1.1
ndg-httpsclient==0.3.2
numpy==1.8.2
oauth2client==4.1.0
picamera==1.13
picraft==1.0
pifacecommon==4.2.1
pifacedigitalio==3.1.0
pigpio==1.35
pyOpenSSL==0.13.1
pyasn1==0.2.3
pyasn1-modules==0.0.9
pycryptodomex==3.4.6
pygame==1.9.2a0
pygobject==3.14.0
pyinotify==0.9.4
pyserial==2.6
python-apt==0.9.3.12
requests==2.4.3
rsa==3.4.2
sense-emu==1.0
sense-hat==2.2.0
six==1.10.0
smbus==1.1
spidev==3.0
twython==3.1.2
uritemplate==3.0.0
urllib3==1.9.1
virtualenv==15.1.0
wheel==0.24.0
wsgiref==0.1.2
